RDK-59014, RDKEMW-7271 : Add support for CURL 56 in urlHelperDownloadFile API and Minor Code Fixes  (#51)
* RDKEMW-7271: Fix GetPartnerId api to return only the value

Reason for change: GetPartnerId was previously returning the entire
key-value pair. This update ensures that only the value is returned
Test Procedure: Verify that GetPartnerId returns the correct value.
Refer to the ticket for detailed steps.
Risks: Low

* RDKEMW-7271: Fix logging format to avoid duplicate newlines

Reason for change: The logging macros previously appended an unnecessary newline ("\n") to the format string, which could result in duplicate newlines or formatting inconsistencies. The macros are updated to pass the format string as-is, relying on the underlying logger to handle line endings appropriately.
Test Procedure: Verify log output. Ensure no extra blank lines appear and log formatting remains consistent.
Risks: Low

* RDK-59014: Add support for CURL 56 in urlHelperDownloadFile API

Reason for change: This update adds handling for CURL error 56 (connection reset by peer) in the urlHelperDownloadFile API. With this change, the seek position is updated appropriately when error 56 occurs, preventing the download from restarting from the beginning.
Test Procedure: Verify that CURL error code 56 is handled in the same way as error codes 18 and 28, and that the seek position is correctly managed.